feat(vscode_ide): Webview code refactor
