+++ ide_extension/vscode/devsecops/src/domain/model/Finding.ts
@@ -10,6 +10,12 @@ export class Finding {
     private description: string;
     private module: string;
     private tool: string;
+    private vulnerability_status?: string;
+    private target_image?: string;
+    private installed_version?: string;
+    private fixed_version?: string;
+    private package_name?: string;
+    private cvss_score?: string;
     private references: string[];
 
     constructor(
@@ -23,6 +29,12 @@ export class Finding {
         description: string,
         module: string,
         tool: string,
+        vulnerability_status?: string,
+        target_image?: string,
+        installed_version?: string,
+        fixed_version?: string,
+        package_name?: string,
+        cvss_score?: string,
         references: string[] = []
     ) {
         this.id = id;
@@ -35,6 +47,12 @@ export class Finding {
         this.description = description;
         this.module = module;
         this.tool = tool;
+        this.vulnerability_status = vulnerability_status;
+        this.target_image = target_image;
+        this.installed_version = installed_version;
+        this.fixed_version = fixed_version;
+        this.package_name = package_name;
+        this.cvss_score = cvss_score;
         this.references = references;
     }
 
@@ -76,6 +94,29 @@ export class Finding {
 
     public getTool(): string {
         return this.tool;
+    }
+
+    public getVulnerabilityStatus(): string | undefined {
+        return this.vulnerability_status;
+    }
+    public getTargetImage(): string | undefined {
+        return this.target_image;
+    }
+
+    public getInstalledVersion(): string | undefined {
+        return this.installed_version;
+    }
+
+    public getFixedVersion(): string | undefined {
+        return this.fixed_version;
+    }
+
+    public getPackageName(): string | undefined {
+        return this.package_name;
+    }
+
+    public getCvssScore(): string | undefined {
+        return this.cvss_score;
     }
 
     public getReferences(): string[] {
